[refactor] useMarkerFeature useEffect 관련 refactor
- hook을 함수처럼 가져다쓰다보니 불필요한 useEffect가 많이 일어남(useEffect의 deps에 redux 상태 포함 때문)
- 위와 같은 불필요한 useEffect 방지를 위해 리팩토링
